AMLS-4703 | moved SectionsProvider to own file
